+++ cymetric/tools.py
@@ -3,9 +3,18 @@
 from __future__ import unicode_literals, print_function
 import os
 import sys
+import warnings
 
 import numpy as np
 import pandas as pd
+
+try:
+    from pyne import data
+    import pyne.enrichment as enr
+    from pyne import nucname
+    HAVE_PYNE = True
+except ImportError:
+    HAVE_PYNE = False
 
 from cyclus import lib
 
@@ -71,6 +80,76 @@ def ensure_dt_bytes(dt):
             ty = ty + t[2:]
         dety.append(ty)
     return dety
+
+
+def format_nuc(nucs):
+    """
+    format the nuclide  provided by the users into a standard format:
+    ZZAASSSS.
+
+    Parameters
+    ----------
+    nucs :  of nuclides
+    """
+    tools.raise_no_pyne('Unable to format nuclide !', HAVE_PYNE)
+
+    return [nucname.id(nuc) for nuc in nucs]
+
+
+def add_missing_time_step(df, ref_time):
+    """
+    Add the missing time step to a Panda Data Frame.
+
+    Parameters
+    ----------
+    df : Pandas Data Frame
+    ref_time :  of the time step references (Coming from TimeStep metrics)
+    """
+    ref_time.rename(index=str, columns={'TimeStep': 'Time'}, inplace=True)
+
+    if 'SimId' in ref_time.columns.values:
+        ref_time.drop('SimId', 1, inplace=True)
+    df = pd.merge(ref_time, df, how="outer")
+    df.fillna(0, inplace=True)
+    return df
+
+
+def merge(df, base_col, add_df, add_col):
+    """
+    Merge some additionnal columns fram an additionnal Pandas Data Frame
+    onother one and then remove the second base column (keeping SimID
+    information).
+
+    Parameters
+    ----------
+    df: Pandas Data Frame
+    base_col:  of the base columns names
+    add_df: Pandas Data Frame to add in the df one
+    add_col: columns to be added
+    """
+    df = pd.merge(add_df[add_col], df, on=base_col)
+    df.drop(base_col[1], 1)
+    return df
+
+
+def reduce(df, rdc_):
+    """
+    Filter the df Pandas Data Frame according to the rdc_ (list of item
+    in the corresponding columns).
+
+    Parameters
+    ----------
+    df: Pandas Data Frame
+    rdc_: list of pair of string and string list.
+    """
+    for rdc in rdc_:
+        if len(rdc[1]) != 0:
+            df = df[df[rdc[0]].isin(rdc[1])]
+        else:
+            wng_msg = "Empty  provided for " + rdc[0] + " key."
+            warnings.warn(wng_msg, UserWarning)
+    return df
+
 
 def raise_no_pyne(msg, have_pyne=False):
     """Raise an error when PyNE cannot be found."""
