+++ tests/test_filters.py
@@ -1,4 +1,4 @@
-"""Tests for convinient interface method"""
+"""Tests for filters method"""
 from __future__ import print_function, unicode_literals
 from uuid import UUID
 import os
@@ -16,14 +16,14 @@ from pandas.util.testing import assert_frame_equal
 from tools import setup, dbtest
 
 import cymetric as cym
-from cymetric import convenient_interface as com
+from cymetric import filters
 from cymetric.tools import raw_to_series, ensure_dt_bytes
 
 
 @dbtest
 def test_convint_get_transaction_df(db, fname, backend):
     evaler = cym.Evaluator(db)
-    cal = com.get_transaction_df(evaler)
+    cal = filters.get_transaction_df(evaler)
     exp_head = ['SimId', 'ReceiverId', 'ReceiverPrototype', 'SenderId',
                 'SenderPrototype', 'TransactionId', 'ResourceId', 'Commodity', 'Time']
     assert_equal(list(cal), exp_head)  # Check we have the correct headers
@@ -52,7 +52,7 @@ def test_convint_get_transaction_df(db, fname, backend):
     assert_frame_equal(cal, refs)
 
     # test single sender
-    cal = com.get_transaction_df(evaler, senders=['UOX_Source'])
+    cal = filters.get_transaction_df(evaler, senders=['UOX_Source'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     # SimId change at each test need to drop it
     cal = cal.drop('TransactionId', 1)
@@ -70,7 +70,7 @@ def test_convint_get_transaction_df(db, fname, backend):
     assert_frame_equal(cal, refs)
 
     # test multiple sender
-    cal = com.get_transaction_df(
+    cal = filters.get_transaction_df(
         evaler, senders=['UOX_Source', 'MOX_Source'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     # SimId change at each test need to drop it
@@ -96,7 +96,7 @@ def test_convint_get_transaction_df(db, fname, backend):
     assert_frame_equal(cal, refs)
 
     # test single receiver
-    cal = com.get_transaction_df(evaler, receivers=['Reactor1'])
+    cal = filters.get_transaction_df(evaler, receivers=['Reactor1'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     # SimId change at each test need to drop it
     cal = cal.drop('TransactionId', 1)
@@ -116,7 +116,7 @@ def test_convint_get_transaction_df(db, fname, backend):
     assert_frame_equal(cal, refs)
 
     # test multiple sender
-    cal = com.get_transaction_df(evaler, receivers=['Reactor1', 'Reactor3'])
+    cal = filters.get_transaction_df(evaler, receivers=['Reactor1', 'Reactor3'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     # SimId change at each test need to drop it
     cal = cal.drop('TransactionId', 1)
@@ -138,7 +138,7 @@ def test_convint_get_transaction_df(db, fname, backend):
     assert_frame_equal(cal, refs)
 
     # test multiple sender and multiple receiver
-    cal = com.get_transaction_df(evaler, senders=['UOX_Source', 'MOX_Source'],
+    cal = filters.get_transaction_df(evaler, senders=['UOX_Source', 'MOX_Source'],
                                  receivers=['Reactor1', 'Reactor2'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     # SimId change at each test need to drop it
@@ -161,8 +161,8 @@ def test_convint_get_transaction_df(db, fname, backend):
     refs.index = refs.index.astype('str')
     assert_frame_equal(cal, refs)
 
-    # test single commodity
-    cal = com.get_transaction_df(evaler, commodities=['uox'])
+    # test single filters.odity
+    cal = filters.get_transaction_df(evaler, commodities=['uox'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     # SimId change at each test need to drop it
     cal = cal.drop('TransactionId', 1)
@@ -180,7 +180,7 @@ def test_convint_get_transaction_df(db, fname, backend):
     assert_frame_equal(cal, refs)
 
     # test multiple sender
-    cal = com.get_transaction_df(evaler, commodities=['uox', 'mox'])
+    cal = filters.get_transaction_df(evaler, commodities=['uox', 'mox'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     # SimId change at each test need to drop it
     cal = cal.drop('TransactionId', 1)
@@ -208,13 +208,13 @@ def test_convint_get_transaction_df(db, fname, backend):
 @dbtest
 def test_convint_get_transaction_nuc_df(db, fname, backend):
     evaler = cym.Evaluator(db)
-    cal = com.get_transaction_nuc_df(evaler)
+    cal = filters.get_transaction_nuc_df(evaler)
     exp_head = ['SimId', 'ResourceId', 'NucId', 'Mass', 'ReceiverId', 'ReceiverPrototype',
                 'SenderId', 'SenderPrototype', 'TransactionId', 'Commodity', 'Time']
     assert_equal(list(cal), exp_head)  # Check we have the correct headers
 
     # test single nuclide selection
-    cal = com.get_transaction_nuc_df(evaler, nucs=['942390000'])
+    cal = filters.get_transaction_nuc_df(evaler, nucs=['942390000'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     # SimId change at each test need to drop it
     cal = cal.drop('TransactionId', 1)
@@ -238,7 +238,7 @@ def test_convint_get_transaction_nuc_df(db, fname, backend):
     assert_frame_equal(cal, refs)
 
     # test multiple nuclide selection
-    cal = com.get_transaction_nuc_df(
+    cal = filters.get_transaction_nuc_df(
         evaler, nucs=['942390000', '922380000'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     # SimId change at each test need to drop it
@@ -275,13 +275,13 @@ def test_convint_get_transaction_nuc_df(db, fname, backend):
 @dbtest
 def test_convint_get_transaction_activity_df(db, fname, backend):
     evaler = cym.Evaluator(db)
-    cal = com.get_transaction_activity_df(evaler)
+    cal = filters.get_transaction_activity_df(evaler)
     exp_head = ['SimId', 'ResourceId', 'NucId', 'Activity', 'ReceiverId', 'ReceiverPrototype',
                 'SenderId', 'SenderPrototype', 'TransactionId', 'Commodity', 'Time']
     assert_equal(list(cal), exp_head)  # Check we have the correct headers
 
     # test single nuclide selection
-    cal = com.get_transaction_activity_df(evaler, nucs=['942390000'])
+    cal = filters.get_transaction_activity_df(evaler, nucs=['942390000'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     # SimId change at each test need to drop it
     cal = cal.drop('TransactionId', 1)
@@ -304,7 +304,7 @@ def test_convint_get_transaction_activity_df(db, fname, backend):
     assert_frame_equal(cal, refs)
 
     # test multiple nuclide selection
-    cal = com.get_transaction_activity_df(
+    cal = filters.get_transaction_activity_df(
         evaler, nucs=['942390000', '922380000'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     # SimId change at each test need to drop it
@@ -341,13 +341,13 @@ def test_convint_get_transaction_activity_df(db, fname, backend):
 @dbtest
 def test_convint_get_transaction_decayheat_df(db, fname, backend):
     evaler = cym.Evaluator(db)
-    cal = com.get_transaction_decayheat_df(evaler)
+    cal = filters.get_transaction_decayheat_df(evaler)
     exp_head = ['SimId', 'ResourceId', 'NucId', 'DecayHeat', 'ReceiverId', 'ReceiverPrototype',
                 'SenderId', 'SenderPrototype', 'TransactionId', 'Commodity', 'Time']
     assert_equal(list(cal), exp_head)  # Check we have the correct headers
 
     # test single nuclide selection
-    cal = com.get_transaction_decayheat_df(evaler, nucs=['942390000'])
+    cal = filters.get_transaction_decayheat_df(evaler, nucs=['942390000'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     # SimId change at each test need to drop it
     cal = cal.drop('TransactionId', 1)
@@ -370,7 +370,7 @@ def test_convint_get_transaction_decayheat_df(db, fname, backend):
     assert_frame_equal(cal, refs)
 
     # test multiple nuclide selection
-    cal = com.get_transaction_decayheat_df(
+    cal = filters.get_transaction_decayheat_df(
         evaler, nucs=['942390000', '922380000'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     # SimId change at each test need to drop it
@@ -404,127 +404,14 @@ def test_convint_get_transaction_decayheat_df(db, fname, backend):
 
 
 @dbtest
-def test_convint_get_transaction_timeseries(db, fname, backend):
-    evaler = cym.Evaluator(db)
-    cal = com.get_transaction_timeseries(evaler)
-    exp_head = ['Time', 'Mass']
-    assert_equal(list(cal), exp_head)  # Check we have the correct headers
-
-    # test single nuclide selection
-    cal = com.get_transaction_timeseries(evaler, nucs=['942390000'])
-    refs = pd.DataFrame(np.array([
-        (0, 0.000000000),
-        (1, 0.0444814879803),
-        (2, 0.0889629759607),
-        (3, 0.0889629759607),
-        (4, 0.0889629759607),
-    ], dtype=ensure_dt_bytes([
-        ('Time', '<i8'), ('Mass', '<f8')
-    ]))
-    )
-    assert_frame_equal(cal, refs)
-
-    # test multiple nuclide selection
-    cal = com.get_transaction_timeseries(
-        evaler, nucs=['942390000', '922380000'])
-    refs = pd.DataFrame(np.array([
-        (0, 0.000000000),
-        (1, 0.831724864011),
-        (2, 1.66344972802),
-        (3, 2.62344972802),
-        (4, 2.62344972802),
-    ], dtype=ensure_dt_bytes([
-        ('Time', '<i8'), ('Mass', '<f8')
-    ]))
-    )
-    assert_frame_equal(cal, refs)
-
-
-@dbtest
-def test_convint_get_transaction_activity_timeseries(db, fname, backend):
-    evaler = cym.Evaluator(db)
-    cal = com.get_transaction_activity_timeseries(evaler)
-    exp_head = ['Time', 'Activity']
-    assert_equal(list(cal), exp_head)  # Check we have the correct headers
-
-    # test single nuclide selection
-    cal = com.get_transaction_activity_timeseries(
-        evaler, nucs=['942390000'])
-    refs = pd.DataFrame(np.array([
-        (0, 0.000000000),
-        (1, 102084984531.0),
-        (2, 204169969062.0),
-        (3, 204169969062.0),
-        (4, 204169969062.0),
-    ], dtype=ensure_dt_bytes([
-        ('Time', '<i8'), ('Activity', '<f8')
-    ]))
-    )
-    assert_frame_equal(cal, refs)
-
-    # test multiple nuclide selection
-    cal = com.get_transaction_activity_timeseries(
-        evaler, nucs=['942390000', '922380000'])
-    refs = pd.DataFrame(np.array([
-        (0, 0.000000000),
-        (1, 102094774891.0),
-        (2, 204189549782.0),
-        (3, 204201488588.0),
-        (4, 204201488588.0),
-    ], dtype=ensure_dt_bytes([
-        ('Time', '<i8'), ('Activity', '<f8')
-    ]))
-    )
-    assert_frame_equal(cal, refs)
-
-
-@dbtest
-def test_convint_get_transaction_decayheat_timeseries(db, fname, backend):
-    evaler = cym.Evaluator(db)
-    cal = com.get_transaction_decayheat_timeseries(evaler)
-    exp_head = ['Time', 'DecayHeat']
-    assert_equal(list(cal), exp_head)  # Check we have the correct headers
-
-    # test single nuclide selection
-    cal = com.get_transaction_decayheat_timeseries(
-        evaler, nucs=['942390000'])
-    refs = pd.DataFrame(np.array([
-        (0, 0.000000000),
-        (1, 3.34065303191e+30),
-        (2, 6.68130606382e+30),
-        (3, 6.68130606382e+30),
-        (4, 6.68130606382e+30),
-    ], dtype=ensure_dt_bytes([
-        ('Time', '<i8'), ('DecayHeat', '<f8')
-    ]))
-    )
-    assert_frame_equal(cal, refs)
-
-    # test multiple nuclide selection
-    cal = com.get_transaction_decayheat_timeseries(
-        evaler, nucs=['942390000', '922380000'])
-    refs = pd.DataFrame(np.array([
-        (0, 0.000000000),
-        (1, 3.34091395721e+30),
-        (2, 6.68182791443e+30),
-        (3, 6.68214609848e+30),
-        (4, 6.68214609848e+30),
-    ], dtype=ensure_dt_bytes([
-        ('Time', '<i8'), ('DecayHeat', '<f8')
-    ]))
-    )
-    assert_frame_equal(cal, refs)
-
-
-@dbtest
 def test_convint_get_inventory_df(db, fname, backend):
     evaler = cym.Evaluator(db)
-    cal = com.get_inventory_df(evaler)
+    cal = filters.get_inventory_df(evaler)
     exp_head = ['SimId', 'AgentId', 'Prototype',
                 'Time', 'InventoryName', 'NucId', 'Quantity']
     assert_equal(list(cal), exp_head)  # Check we have the correct headers
 
-    cal = com.get_inventory_df(evaler, facilities=['Reactor1'],
+    cal = filters.get_inventory_df(evaler, facilities=['Reactor1'],
                                nucs=['94239'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     refs = pd.DataFrame(np.array([
@@ -541,7 +428,7 @@ def test_convint_get_inventory_df(db, fname, backend):
     )
     assert_frame_equal(cal, refs)
 
-    cal = com.get_inventory_df(evaler, facilities=['Reactor1'],
+    cal = filters.get_inventory_df(evaler, facilities=['Reactor1'],
                                nucs=['94239', '92235'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     refs = pd.DataFrame(np.array([
@@ -569,12 +456,12 @@ def test_convint_get_inventory_df(db, fname, backend):
 @dbtest
 def test_convint_get_inventory_activity_df(db, fname, backend):
     evaler = cym.Evaluator(db)
-    cal = com.get_inventory_activity_df(evaler)
+    cal = filters.get_inventory_activity_df(evaler)
     exp_head = ['SimId', 'AgentId', 'Prototype', 'Time', 'InventoryName',
             'NucId', 'Quantity', 'Activity']
     assert_equal(list(cal), exp_head)  # Check we have the correct headers
 
-    cal = com.get_inventory_activity_df(evaler, facilities=['Reactor1'],
+    cal = filters.get_inventory_activity_df(evaler, facilities=['Reactor1'],
                                nucs=['94239'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     refs = pd.DataFrame(np.array([
@@ -597,7 +484,7 @@ def test_convint_get_inventory_activity_df(db, fname, backend):
     )
     assert_frame_equal(cal, refs)
 
-    cal = com.get_inventory_activity_df(evaler, facilities=['Reactor1'],
+    cal = filters.get_inventory_activity_df(evaler, facilities=['Reactor1'],
                                nucs=['94239', '92235'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     refs = pd.DataFrame(np.array([
@@ -631,12 +518,12 @@ def test_convint_get_inventory_activity_df(db, fname, backend):
 @dbtest
 def test_convint_get_inventory_decayheat_df(db, fname, backend):
     evaler = cym.Evaluator(db)
-    cal = com.get_inventory_decayheat_df(evaler)
+    cal = filters.get_inventory_decayheat_df(evaler)
     exp_head = ['SimId', 'AgentId', 'Prototype', 'Time', 'InventoryName',
             'NucId', 'Quantity', 'Activity', 'DecayHeat']
     assert_equal(list(cal), exp_head)  # Check we have the correct headers
 
-    cal = com.get_inventory_decayheat_df(evaler, facilities=['Reactor1'],
+    cal = filters.get_inventory_decayheat_df(evaler, facilities=['Reactor1'],
                                nucs=['94239'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     refs = pd.DataFrame(np.array([
@@ -660,7 +547,7 @@ def test_convint_get_inventory_decayheat_df(db, fname, backend):
     )
     assert_frame_equal(cal, refs)
 
-    cal = com.get_inventory_decayheat_df(evaler, facilities=['Reactor1'],
+    cal = filters.get_inventory_decayheat_df(evaler, facilities=['Reactor1'],
                                nucs=['94239', '92235'])
     cal = cal.drop('SimId', 1)  # SimId change at each test need to drop it
     refs = pd.DataFrame(np.array([
@@ -694,114 +581,6 @@ def test_convint_get_inventory_decayheat_df(db, fname, backend):
         ('AgentId', '<i8'), ('Prototype', 'O'), ('Time', '<i8'),
         ('InventoryName', 'O'), ('NucId', '<i8'), ('Quantity', '<f8'),
         ('Activity', '<f8'), ('DecayHeat', '<f8')
-    ]))
-    )
-    assert_frame_equal(cal, refs)
-
-
-@dbtest
-def test_convint_get_inventory_timeseries(db, fname, backend):
-    evaler = cym.Evaluator(db)
-    cal = com.get_inventory_timeseries(evaler)
-    exp_head = ['Time', 'Quantity']
-    assert_equal(list(cal), exp_head)  # Check we have the correct headers
-
-    cal = com.get_inventory_timeseries(evaler, facilities=['Reactor1'],
-                               nucs=['94239'])
-    refs = pd.DataFrame(np.array([
-        (0, 0.0            ),
-        (1, 0.0444814879803),
-        (2, 0.0621806030246),
-        (3, 0.0798797180688),
-        (4, 0.0530973451327)
-    ], dtype=ensure_dt_bytes([
-        ('Time', '<i8'), ('Quantity', '<f8')
-    ]))
-    )
-    assert_frame_equal(cal, refs)
-
-    cal=com.get_inventory_timeseries(evaler, facilities=['Reactor1'],
-                               nucs=['94239', '92235'])
-    refs=pd.DataFrame(np.array([
-        (0, 0.0            ),
-        (1, 0.0460607124057),
-        (2, 0.0726093849721),
-        (3, 0.0991580575384),
-        (4, 0.119646017699 )
-    ], dtype=ensure_dt_bytes([
-        ('Time', '<i8'), ('Quantity', '<f8')
-    ]))
-    )
-    assert_frame_equal(cal, refs)
-
-
-@dbtest
-def test_convint_get_inventory_activity_timeseries(db, fname, backend):
-    evaler=cym.Evaluator(db)
-    cal=com.get_inventory_activity_timeseries(evaler)
-    exp_head=['Time', 'Activity']
-    assert_equal(list(cal), exp_head)  # Check we have the correct headers
-
-    cal=com.get_inventory_activity_timeseries(evaler, facilities=['Reactor1'],
-                               nucs=['94239'])
-    refs=pd.DataFrame(np.array([
-        (0, 0.0             ),
-        (1, 2.44036364223e+13),
-        (2, 3.41138054869e+13),
-        (3, 4.38239745515e+13),
-        (4, 2.91305071939e+13)
-    ], dtype=ensure_dt_bytes([
-        ('Time', '<i8'), ('Activity', '<f8')
-    ]))
-    )
-    assert_frame_equal(cal, refs)
-
-    cal=com.get_inventory_activity_timeseries(evaler, facilities=['Reactor1'],
-                               nucs=['94239', '92235'])
-    refs=pd.DataFrame(np.array([
-        (0, 0.0             ),
-        (1, 2.4403666094e+13),
-        (2, 3.41140014315e+13),
-        (3, 4.3824336769e+13),
-        (4, 2.91317575657e+13),
-    ], dtype=ensure_dt_bytes([
-        ('Time', '<i8'), ('Activity', '<f8')
-    ]))
-    )
-    assert_frame_equal(cal, refs)
-
-
-@dbtest
-def test_convint_get_inventory_decayheat_timeseries(db, fname, backend):
-    evaler=cym.Evaluator(db)
-    cal=com.get_inventory_decayheat_timeseries(evaler)
-    exp_head=['Time', 'DecayHeat']
-    assert_equal(list(cal), exp_head)  # Check we have the correct headers
-
-    cal=com.get_inventory_decayheat_timeseries(evaler, facilities=['Reactor1'],
-                               nucs=['94239'])
-    refs=pd.DataFrame(np.array([
-        (0, 0.0             ),
-        (1, 7.98590335085e+32),
-        (2, 1.11634819022e+33),
-        (3, 1.43410604536e+33),
-        (4, 9.53273565408e+32)
-    ], dtype=ensure_dt_bytes([
-        ('Time', '<i8'), ('DecayHeat', '<f8')
-    ]))
-    )
-    assert_frame_equal(cal, refs)
-
-    cal=com.get_inventory_decayheat_timeseries(evaler, facilities=['Reactor1'],
-                               nucs=['94239', '92235'])
-    refs=pd.DataFrame(np.array([
-        (0, 0.0             ),
-        (1, 7.98591200694e+32),
-        (2, 1.11635390648e+33),
-        (3, 1.43411661226e+33),
-        (4, 9.53310042276e+32)
-    ], dtype=ensure_dt_bytes([
-        ('Time', '<i8'), ('DecayHeat', '<f8')
     ]))
     )
     assert_frame_equal(cal, refs)
